Restore original creation date
